cameera: scx15: Fix compiler warnings
This is probably the largest codebase from SPRD and also "the"
nightmare when it comes to the  -Werror change from android 9.

It's difficult and wasteful to modify the source code just to iron
out ANY warnings so I started by ignoring these set of warnings:
unused-variable
unused-parameter
unused-function (only for the utest)

Then we set on fixing the rest of the warnings which makes sense.
An example of a warning that makes sense is the warning that appear
on a wrong format specifier in printf or even in ALOG functions, also
including too few or too many specifiers compared to the arguments passed.

+++ libcamera/scx15/isp/isp_param_file_update.c
@@ -200,7 +200,7 @@ uint32_t isp_raw_para_update_from_file(SENSOR_INFO_T *sensor_info_ptr,SENSOR_ID_
 		CMR_LOGE("file first load");
 		raw_info_update_status[sensor_id].tune_info_table_org_addr = (uint8_t*)(*sensor_info_ptr->raw_info_ptr)->tune_ptr;
 		raw_info_update_status[sensor_id].fix_info_table_org_addr = (*sensor_info_ptr->raw_info_ptr)->fix_ptr;
-	}else if(raw_info_update_status[sensor_id].updata_file_time == file_status.st_mtime){
+	}else if(raw_info_update_status[sensor_id].updata_file_time == (unsigned long) file_status.st_mtime){
 		CMR_LOGE("file has been loaded");
 		if(fp){
 			fclose(fp);
